ui: move log components inside system-logs dir
